errata module - moving it from controllers to lib
