+++ src/app/controllers/system_errata_controller.rb
@@ -10,6 +10,8 @@
 # have received a copy of GPLv2 along with this software; if not, see
 # http://www.gnu.org/licenses/old-licenses/gpl-2.0.txt.
 
+require 'util/errata'
+
 class SystemErrataController < ApplicationController
 
   before_filter :find_system, :only =>[:install, :index, :items, :status]
@@ -84,7 +86,7 @@ class SystemErrataController < ApplicationController
   private
 
   include SortColumnList
-  include ErrataModule
+  include Katello::Errata
 
   def get_errata start, finish, filter_type="All", errata_state="outstanding"
     types = [Glue::Pulp::Errata::SECURITY, Glue::Pulp::Errata::ENHANCEMENT, Glue::Pulp::Errata::BUGZILLA]
